requirements.txt
discord.py[voice]
gtts
